api/requirements.txt
asgiref==3.5.2
certifi==2022.6.15.2
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
defusedxml==0.7.1
dj-rest-auth==2.2.5
Django==4.1.1
django-allauth==0.50.0
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-camel-case==1.3.0
djangorestframework-simplejwt==5.2.0
drf-access-policy==1.3.0
drf-extensions==0.7.1
idna==3.4
importlib-metadata==4.12.0
Markdown==3.4.1
oauthlib==3.2.1
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
sqlparse==0.4.2
tzdata==2022.2
urllib3==1.26.12
whitenoise==6.2.0
zipp==3.8.1
